pyproject.toml
[tool.poetry]
name = "botadminbackend"
version = "0.1.0"
description = ""
authors = ["Akulij <aakulij@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.97.0"}
sqlmodel = "^0.0.8"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
alembic = "^1.11.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
